Bugfix on PickUpAndExamine Skript
Due to this Script not being made by me and given to us by our Professor, i modified it to use the StarterAssetInput-System.
After A while we stumbled upon a bug, where the player was able to drop the item while examining, locking him into a state of not being able to move and being forced to restart the game.
With the help of ChatGPT i was able to fix this issue.